LayoutTests/fast/selectors/lang-empty-expected.txt

Verify empty :lang() selectors.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.querySelectorAll(":lang()").length threw exception SyntaxError: ':lang()' is not a valid selector..
PASS document.querySelectorAll(":lang() *").length threw exception SyntaxError: ':lang() *' is not a valid selector..
PASS getComputedStyle(document.getElementById("target1")).color is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
